Automatically backed up by Learn

+++ db/migrate/001_create_costumes.rb
@@ -12,7 +12,7 @@ class CreateCostumes < ActiveRecord::Migration[5.1]
   def change
     create_table :costumes do |x|
       x.string :name 
-      x.decimal :price 
+      x.float :price 
       x.string :size 
       x.string :image_url
       x.timestamps
